+++ src/wrap/append.js
@@ -1,1 +1,6 @@
+
+// AMD Compliance
+if ('function' === typeof define && define.amd)
+  define('parsley', function() { return window.Parsley; } );
+
 })(window.jQuery);